+++ src/_internal/MakeBom/Options.php
@@ -100,9 +100,9 @@ class Options
      * @psalm-var array<string, Version>
      */
     private const VALUE_SPEC_VERSION_MAP = [
-        '1.4' => Version::v1dot4,
-        // first in list is the default value - see constructor
         '1.5' => Version::v1dot5,
+        // first in list is the default value - see constructor
+        '1.4' => Version::v1dot4,
         '1.3' => Version::v1dot3,
         '1.2' => Version::v1dot2,
         '1.1' => Version::v1dot1,